requirements.txt
anthropic>=0.25.0
sentence-transformers>=2.7.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
